Undoing API-compatibility breaking changes with upstream.
Yes, errs is not touched at all in mbe_process*Dataf(). Should probably talk to upstream about that, though again, probably guess they don't want to break API either.

Probably the best idea is cramming a bunch of those arguments into a struct, defining a new function that uses that, and just using that in all of my code (with a compat wrapper if requested by anyone).